funzionalità tavolo e admin terminate

+++ src/main/java/it/pokeronline/model/tavolo/Tavolo.java
@@ -22,7 +22,7 @@ public class Tavolo {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private Long expMin;
-	private Double cifraMin;
+	private Integer cifraMin;
 	private String denominazione;
 	
 	@Temporal(TemporalType.DATE)
@@ -39,7 +39,7 @@ public class Tavolo {
 	public Tavolo(Long id) {
 		this.id = id;
 	}
-	public Tavolo(Long expMin, Double cifraMin, String denominazione, Date dataCreazione) {
+	public Tavolo(Long expMin, Integer cifraMin, String denominazione, Date dataCreazione) {
 		super();
 		this.expMin = expMin;
 		this.cifraMin = cifraMin;
@@ -47,7 +47,7 @@ public class Tavolo {
 		this.dataCreazione = dataCreazione;
 	}
 
-	public Tavolo(Double cifraMin, String denominazione, Date date) {
+	public Tavolo(Integer cifraMin, String denominazione, Date date) {
 	 this.cifraMin = cifraMin;
 	 this.denominazione = denominazione;
 	 this.dataCreazione = date;
@@ -68,11 +68,11 @@ public class Tavolo {
 		this.expMin = expMin;
 	}
 
-	public Double getCifraMin() {
+	public Integer getCifraMin() {
 		return cifraMin;
 	}
 
-	public void setCifraMin(Double cifraMin) {
+	public void setCifraMin(Integer cifraMin) {
 		this.cifraMin = cifraMin;
 	}
